pipenv/synapse/requirements.txt
pandas~=2.2.3
pyarrow~=17.0.0
mysql-connector-python~=9.0.0
SQLAlchemy~=2.0.35